# Conflicts:
#	src/main/java/com/beehyv/tbalert/tbalertbackend/controller/ContactScreeningController.java
#	src/main/java/com/beehyv/tbalert/tbalertbackend/controller/PatientController.java
#	src/main/java/com/beehyv/tbalert/tbalertbackend/controller/SettingController.java
#	src/main/java/com/beehyv/tbalert/tbalertbackend/controller/StateHeadController.java
#	src/main/java/com/beehyv/tbalert/tbalertbackend/controller/TeleCallerController.java
#	src/main/java/com/beehyv/tbalert/tbalertbackend/entity/Setting.java
#	src/main/java/com/beehyv/tbalert/tbalertbackend/mapper/StateHeadMapper.java
#	src/main/java/com/beehyv/tbalert/tbalertbackend/repository/MissedMedicationRepo.java
#	src/main/java/com/beehyv/tbalert/tbalertbackend/repository/PatientFollowUpRepo.java
#	src/main/java/com/beehyv/tbalert/tbalertbackend/repository/PersonRepo.java
#	src/main/java/com/beehyv/tbalert/tbalertbackend/repository/StateHeadRepo.java
#	src/main/java/com/beehyv/tbalert/tbalertbackend/repository/TeleCallerRepo.java
#	src/main/java/com/beehyv/tbalert/tbalertbackend/service/StateHeadService.java
#	src/main/java/com/beehyv/tbalert/tbalertbackend/service/TeleCallerService.java
#	src/main/java/com/beehyv/tbalert/tbalertbackend/service/impl/PatientMedicationServiceImpl.java
#	src/main/java/com/beehyv/tbalert/tbalertbackend/service/impl/PatientRegistrationServiceImpl.java
#	src/main/java/com/beehyv/tbalert/tbalertbackend/service/impl/PersonServiceImpl.java
#	src/main/java/com/beehyv/tbalert/tbalertbackend/service/impl/TeleCallerServiceImpl.java
#	src/main/java/com/beehyv/tbalert/tbalertbackend/specifications/PatientSpecification.java

+++ src/main/java/com/beehyv/tbalert/tbalertbackend/entity/Setting.java
@@ -1,11 +1,6 @@
 package com.beehyv.tbalert.tbalertbackend.entity;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -20,8 +15,7 @@ import lombok.NoArgsConstructor;
 public class Setting {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
-
+    private Long id;
     @Column
     private String keyName;
